userservice/management/requirements.txt
Django==5.0.3
djangorestframework==3.15.0
PyJWT==2.8.0
PyJWT==2.8.0
razorpay==1.4.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
gunicorn==21.2.0
